feat: add Footer component
