+++ src/pages/layouts/BaseLayout/BaseLayout.module.css
@@ -1,0 +1,13 @@
+.main {
+  box-sizing: border-box;
+  max-width: var(--sizes-maxContent);
+  margin: 2.25rem auto 0px;
+  padding-left: 1.5rem;
+  padding-right: 1.5rem;
+}
+
+@media (min-width: 576px) {
+  .main {
+    margin-top: 2.5rem;
+  }
+}
