+++ src/pages/layouts/BaseLayout/components/AppFooter/AppFooter.module.css
@@ -1,0 +1,118 @@
+.footer {
+  margin-top: 7.5rem;
+  border-top: 1px solid var(--gray-scale-100);
+  padding-top: 2rem;
+  padding-bottom: 6rem;
+}
+
+.footer h4 {
+  font-size: var(--fontSizes-caption2);
+}
+
+.footer a:hover {
+  opacity: 0.64;
+}
+
+.footer-container {
+  display: grid;
+  box-sizing: border-box;
+  max-width: var(--sizes-maxContent);
+  margin: 0 auto;
+  padding-left: 1.5rem;
+  padding-right: 1.5rem;
+  width: 100%;
+  display: grid;
+  gap: 2.625rem;
+}
+
+.nav-container {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.nav-left {
+  display: flex;
+  flex-direction: column;
+  gap: 1.75rem;
+  /* display: flex;
+  gap: 3rem; */
+}
+.nav-left > li:first-of-type > a > h4 {
+  font-weight: bold;
+}
+
+.nav-right {
+  margin-top: 1.75rem;
+  display: flex;
+  gap: 1.5rem;
+}
+
+.external-link {
+  display: inline-flex;
+  width: 2rem;
+  height: 2rem;
+  align-items: center;
+  justify-content: center;
+}
+.external-link > a > svg {
+  width: 1em;
+  height: 1em;
+  color: var(--gray-scale-900);
+}
+.external-link > a {
+  display: inline-block;
+  font-size: 1.5rem;
+  line-height: 0;
+}
+.external-link > a:hover {
+  opacity: 0.64;
+}
+
+.address-container {
+  display: flex;
+  gap: 1rem;
+  flex-direction: column;
+  color: var(--gray-scale-600);
+  font-size: var(--fontSizes-caption1);
+}
+
+.address {
+  display: flex;
+  max-width: 37.5rem;
+  flex-wrap: wrap;
+  gap: 0.25rem 0.5rem;
+}
+
+.address a {
+  color: var(--seed-scale-color-gray-600);
+  text-decoration: underline;
+}
+
+@media (min-width: 576px) {
+  .footer {
+    padding-top: 4.375rem;
+    margin-top: 10rem;
+  }
+  .footer-container {
+    gap: 2.25rem;
+  }
+}
+
+@media (min-width: 768px) {
+  .nav-container {
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    align-items: center;
+    justify-content: space-between;
+  }
+  .nav-left {
+    flex-direction: row;
+    gap: 3rem;
+  }
+
+  .nav-right {
+    margin-top: 0px;
+  }
+}
