+++ src/pages/layouts/BaseLayout/components/AppHeader/AppHeader.module.css
@@ -86,7 +86,8 @@ label.menu-toggle {
   width: 24px;
   height: 24px;
 }
-.stroke1 {
+
+.stroke {
   fill: none;
   stroke: var(--gray-scale-900);
   stroke-width: 6;
@@ -95,20 +96,7 @@ label.menu-toggle {
   transition: stroke-dasharray 600ms cubic-bezier(0.4, 0, 0.2, 1),
     stroke-dashoffset 600ms cubic-bezier(0.4, 0, 0.2, 1);
 }
-.input:checked + label .stroke1 {
-  stroke-dasharray: 90 207;
-  stroke-dashoffset: -134;
-}
-.stroke2 {
-  fill: none;
-  stroke: var(--gray-scale-900);
-  stroke-width: 6;
-  stroke-linecap: round;
-  stroke-dasharray: 60 207;
-  transition: stroke-dasharray 600ms cubic-bezier(0.4, 0, 0.2, 1),
-    stroke-dashoffset 600ms cubic-bezier(0.4, 0, 0.2, 1);
-}
-.input:checked + label .stroke2 {
+.input:checked + label .stroke {
   stroke-dasharray: 90 207;
   stroke-dashoffset: -134;
 }
